Add Digits component inside calculator component, Move App.js to src folder from components folder.
